mastering-carol/data-app-bq-ml/requirements.txt
geopandas==0.13.2
scikit-learn==1.3.2
matplotlib==3.8.2
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
pycarol==2.55.0
db-dtypes==1.1.1
